+++ diesel_cli/src/infer_schema_internals/sqlite.rs
@@ -5,6 +5,7 @@ use diesel::*;
 
 use super::data_structures::*;
 use super::table_data::TableName;
+use crate::print_schema::ColumnSorting;
 
 table! {
     sqlite_master (name) {
@@ -91,9 +92,19 @@ pub fn load_foreign_key_constraints(
 pub fn get_table_data(
     conn: &SqliteConnection,
     table: &TableName,
+    column_sorting: &ColumnSorting,
 ) -> QueryResult<Vec<ColumnInformation>> {
     let query = format!("PRAGMA TABLE_INFO('{}')", &table.sql_name);
-    sql::<pragma_table_info::SqlType>(&query).load(conn)
+    let mut result = sql::<pragma_table_info::SqlType>(&query).load(conn)?;
+    match column_sorting {
+        ColumnSorting::OrdinalPosition => {}
+        ColumnSorting::Name => {
+            result.sort_by(|a: &ColumnInformation, b: &ColumnInformation| {
+                a.column_name.partial_cmp(&b.column_name).unwrap()
+            });
+        }
+    };
+    Ok(result)
 }
 
 #[derive(Queryable)]