+++ diesel_cli/tests/print_schema/print_schema_column_order/postgres/expected.rs
@@ -1,0 +1,9 @@
+// @generated automatically by Diesel CLI.
+
+diesel::table! {
+    abc (a) {
+        a -> Int4,
+        b -> Varchar,
+        c -> Nullable<Bool>,
+    }
+}