+++ diesel_cli/tests/print_schema/print_schema_column_order/sqlite/expected.rs
@@ -1,0 +1,9 @@
+// @generated automatically by Diesel CLI.
+
+diesel::table! {
+    abc (a) {
+        a -> Nullable<Integer>,
+        b -> Text,
+        c -> Nullable<Bool>,
+    }
+}